Fix dtype

+++ scico/examples.py
@@ -666,7 +666,7 @@ def create_laminar_phantom() -> np.ndarray:
     Returns:
         Laminar phantom.
     """
-    vol = np.zeros((64, 256, 256))
+    vol = np.zeros((64, 256, 256), dtype=np.float32)
 
     layer1 = np.zeros(vol.shape[1:])
     layer1blks = _extract_blocks(layer1, (48, 48), (64, 64))
